verticals/requirements.txt
asgiref==3.2.5
dj-database-url==0.5.0
Django==3.0.4
pytz==2019.3
sqlparse==0.3.1
django-filter==2.2.0
djangorestframework==3.11.0
Pillow==7.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.3